// src/generated/client/models/ChangeProposalStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Lipsanen Project Management API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Change proposal status
 * @export
 */
export const ChangeProposalStatus = {
    Pending: 'PENDING',
    Approved: 'APPROVED',
    Rejected: 'REJECTED'
} as const;
export type ChangeProposalStatus = typeof ChangeProposalStatus[keyof typeof ChangeProposalStatus];


export function ChangeProposalStatusFromJSON(json: any): ChangeProposalStatus {
    return ChangeProposalStatusFromJSONTyped(json, false);
}

export function ChangeProposalStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChangeProposalStatus {
    return json as ChangeProposalStatus;
}

export function ChangeProposalStatusToJSON(value?: ChangeProposalStatus | null): any {
    return value as any;
}